fix chained commands proceeding even on error
Regression introduced in b92c757: the atomic bool was erroneously
replaced with closing the cmdResult channel. When the cmdResult channel
is closed, subsequent reads will keep yielding the zero value (a nil
error in this case). Therefore failed commands which should rightfully
break the chain end up proceeding anyway because first it fails, then it
"succeeds" (because select{} on the closed channel produces a zero
value). Fixed by introducing an additional channel waitDone which is
used to actually signal when a command is done.