gracefully handle too many open files instead of crashing
When we hit "too many open files" (syscall.EMFILE), unwatch the last 256 files
(or 20% of the currently watched files, whichever is smaller) so that the
commands after that may use their own file descriptors and not have all file
descriptors sucked up by the watcher. In the future we may decide to gracefully
fallback to polling, but this error really only happens when somebody does
something crazy like watching their entire home directory and I'm not sure I
want to empower that.