fix bug where we were unwatching everything but 256 files, instead of just unwatching 256 files
Unwatching occurs when we hit the syscall.EMFILE error ("too many open files"),
and we need to unwatch so that we have some file descriptors available for
starting commands i.e. cmd.Start(). On macOS, creating a timer for debouncing
already consumes a file descriptor [1], and if there are no file descriptors
available it panics with a very intimidating stack trace.

[1] https://dzone.com/articles/go-servers-understanding-epoll-kqueue-netpoll#:~:text=kqueue%20is%20a%20similar%20mechanism%2C%20where%20registered%20events%20(read/write%2C%20timers%2C%20signals%2C%20files)%20are%20handled%20via%20kevent